pyproject.toml
[tool.poetry]
name = "fastapi-playtime"
version = "0.1.0"
description = ""
authors = [
    "Everson Esteves <esteveseverson@gmail.com>",
    "Gabriela Fernandes <gabiosfc@gmail.com>"
]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12.*"
fastapi = {extras = ["standart"], version = "^0.115.5"}
sqlalchemy = "^2.0.36"
alembic = "^1.14.0"
pwdlib = {extras = ["argon2"], version = "^0.2.1"}
pydantic-settings = "^2.6.1"
pyjwt = "^2.10.0"
tzdata = "^2024.2"
python-multipart = "^0.0.17"

[tool.poetry.group.dev.dependencies]
ruff = "^0.8.0"
pytest = "^8.3.3"
pytest-cov = "^6.0.0"
taskipy = "^1.14.0"

[tool.pytest.ini_options]
pythonpath = "."
addopts = '-p no:warnings'

[tool.ruff]
line-length = 79
extend-exclude = ['migrations']

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']
ignore = ['PLR']

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.taskipy.tasks]
run = 'fastapi dev fastapi_playtime/app.py'
format = 'ruff check . --fix && ruff format .'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
